move ui packages into root folder
