+++ src/main/java/de/marhali/easyi18n/action/treeview/CollapseTreeViewAction.java
@@ -1,4 +1,4 @@
-package de.marhali.easyi18n.ui.action.treeview;
+package de.marhali.easyi18n.action.treeview;
 
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnAction;
